+++ spring-boot-example/src/main/java/io/github/manuelarte/spring/queryparameter/mongo/example/repositories/Example1Repository.java
@@ -1,0 +1,10 @@
+package io.github.manuelarte.spring.queryparameter.mongo.example.repositories;
+
+
+import io.github.manuelarte.spring.queryparameter.mongo.example.entities.Example1;
+import org.bson.types.ObjectId;
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+public interface Example1Repository extends MongoRepository<Example1, ObjectId> {
+
+}