+++ src/main/java/io/github/manuelarte/spring/queryparameter/mongo/operatorcriteria/DefaultGreaterThanCriteria.java
@@ -4,7 +4,7 @@ import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.stereotype.Component;
 
 @Component
-public class DefaultGreaterThanCriteria<V> implements OperatorCriteria<V> {
+public class DefaultGreaterThanCriteria<V> implements OperatorQueryCriteria<V> {
 
   @Override
   public Criteria apply(final String key, final V value) {