+++ src/main/java/io/github/manuelarte/spring/queryparameter/mongo/operatorcriteria/DefaultNotInCriteria.java
@@ -1,14 +1,16 @@
 package io.github.manuelarte.spring.queryparameter.mongo.operatorcriteria;
 
+import java.util.Collection;
 import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.stereotype.Component;
 
 @Component
-public class DefaultLowerThanOrEqualsCriteria<V> implements OperatorQueryCriteria<V> {
+public class DefaultNotInCriteria<V> implements OperatorQueryCriteria<V> {
 
   @Override
   public Criteria apply(final String key, final V value) {
-    return Criteria.where(key).lte(value);
+    return Criteria.where(key).not().in((Collection) value);
   }
+
 
 }