+++ src/main/java/io/github/manuelarte/spring/queryparameter/mongo/operatorcriteria/DefaultNotLowerThanCriteria.java
@@ -4,11 +4,11 @@ import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.stereotype.Component;
 
 @Component
-public class DefaultLowerThanOrEqualsCriteria<V> implements OperatorQueryCriteria<V> {
+public class DefaultNotLowerThanCriteria<V> implements OperatorQueryCriteria<V> {
 
   @Override
   public Criteria apply(final String key, final V value) {
-    return Criteria.where(key).lte(value);
+    return Criteria.where(key).not().lt(value);
   }
 
 }