+++ src/main/java/io/github/manuelarte/spring/queryparameter/mongo/transformers/QueryCriteriaToMongoQueryTransformerImpl.java
@@ -1,13 +1,14 @@
 package io.github.manuelarte.spring.queryparameter.mongo.transformers;
 
 import io.github.manuelarte.spring.queryparameter.model.TypeTransformerProvider;
-import io.github.manuelarte.spring.queryparameter.mongo.model.OperatorCriteriaProvider;
-import io.github.manuelarte.spring.queryparameter.mongo.operatorcriteria.OperatorCriteria;
+import io.github.manuelarte.spring.queryparameter.mongo.operatorcriteria.OperatorQueryCriteria;
 import io.github.manuelarte.spring.queryparameter.operators.Operator;
+import io.github.manuelarte.spring.queryparameter.operators.queryprovider.OperatorQueryProvider;
 import io.github.manuelarte.spring.queryparameter.query.BooleanOperator;
 import io.github.manuelarte.spring.queryparameter.query.OtherCriteria;
 import io.github.manuelarte.spring.queryparameter.query.QueryCriteria;
 import io.github.manuelarte.spring.queryparameter.query.QueryCriterion;
+import io.github.manuelarte.spring.queryparameter.transformers.QueryCriteriaTransformer;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -17,16 +18,21 @@ import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 
 public class QueryCriteriaToMongoQueryTransformerImpl implements
-    QueryCriteriaToMongoQueryTransformer {
+    QueryCriteriaTransformer<Query> {
 
   private final TypeTransformerProvider typeTransformerProvider;
-  private final OperatorCriteriaProvider operatorCriteriaProvider;
+  private final OperatorQueryProvider<OperatorQueryCriteria<Object>, Criteria> operatorCriteriaProvider;
 
   public QueryCriteriaToMongoQueryTransformerImpl(
       final TypeTransformerProvider typeTransformerProvider,
-      final OperatorCriteriaProvider operatorCriteriaProvider) {
+      final OperatorQueryProvider<OperatorQueryCriteria<Object>, Criteria> operatorCriteriaProvider) {
     this.typeTransformerProvider = typeTransformerProvider;
     this.operatorCriteriaProvider = operatorCriteriaProvider;
+  }
+
+  @Override
+  public boolean supports(final Class<?> clazz) {
+    return Query.class.equals(clazz);
   }
 
   @Override
@@ -34,12 +40,12 @@ public class QueryCriteriaToMongoQueryTransformerImpl implements
     if (queryCriteria == null) {
       return null;
     } else {
-      final MultiValueMap<BooleanOperator, QueryCriterion> grouped = groupByBooleanOperator(
+      final MultiValueMap<BooleanOperator, QueryCriterion<Object>> grouped = groupByBooleanOperator(
           queryCriteria);
       final Criteria criteria = new Criteria();
       grouped.keySet().forEach(qO -> {
         Criteria[] qOCriteria = grouped.get(qO).stream().map(it -> createCriteria(entity, it))
-            .toArray(size -> new Criteria[size]);
+            .toArray(Criteria[]::new);
         switch (qO) {
           case AND:
             criteria.andOperator(qOCriteria);
@@ -55,9 +61,9 @@ public class QueryCriteriaToMongoQueryTransformerImpl implements
     }
   }
 
-  private MultiValueMap<BooleanOperator, QueryCriterion> groupByBooleanOperator(
+  private MultiValueMap<BooleanOperator, QueryCriterion<Object>> groupByBooleanOperator(
       final QueryCriteria queryCriteria) {
-    final MultiValueMap<BooleanOperator, QueryCriterion> multiValueMap =
+    final MultiValueMap<BooleanOperator, QueryCriterion<Object>> multiValueMap =
         new LinkedMultiValueMap<>();
     // add first query criterion
     multiValueMap.add(BooleanOperator.AND, queryCriteria.getCriterion());
@@ -65,14 +71,14 @@ public class QueryCriteriaToMongoQueryTransformerImpl implements
     while (otherCriteriaOptional.isPresent()) {
       final OtherCriteria other = otherCriteriaOptional.get();
       final QueryCriteria newQueryCriteria = other.getCriteria();
-      final QueryCriterion newQueryCriterion = newQueryCriteria.getCriterion();
+      final QueryCriterion<Object> newQueryCriterion = newQueryCriteria.getCriterion();
       multiValueMap.add(other.getOperator(), newQueryCriterion);
       otherCriteriaOptional = newQueryCriteria.getOther();
     }
     return multiValueMap;
   }
 
-  private Criteria createCriteria(final Class<?> entity, final QueryCriterion queryCriterion) {
+  private Criteria createCriteria(final Class<?> entity, final QueryCriterion<Object> queryCriterion) {
     final Object castedValue;
     if (queryCriterion.getValue() instanceof List) {
       castedValue = ((List) queryCriterion.getValue()).stream()
@@ -84,8 +90,8 @@ public class QueryCriteriaToMongoQueryTransformerImpl implements
           queryCriterion.getKey()).transformValue(entity, queryCriterion.getKey(),
           queryCriterion.getValue());
     }
-    final OperatorCriteria<Object> operatorPredicate = operatorCriteriaProvider
-        .getOperatorCriteria(entity, queryCriterion.getKey(),
+    final OperatorQueryCriteria<Object> operatorPredicate = operatorCriteriaProvider
+        .getOperatorQuery(entity, queryCriterion.getKey(),
             (Operator<Object>) queryCriterion.getOperator());
     return operatorPredicate.apply(queryCriterion.getKey(), castedValue);
   }