+++ CHANGELOG.md
@@ -4,9 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2] - 2020-04-23
+
+### Added
+
+- Refactor to the new model 0.0.2 changes.
+- Added support for not operators.
+
 ## [0.0.1] - 2020-04-20
 
 ### Added
 
-- Added support for equals, lower than, lower than or equal, greater than, greater than or equals, in operators
+- Added support for equals, lower than, lower than or equal, greater than, greather than or equals, in operators
 - Added support to allow or not allow certain keys.